Extract ClassInstance
