+++ lib/toy/class_instance.rb
@@ -1,0 +1,20 @@
+require "toy/module_instance"
+require "toy/object_instance"
+
+module Toy
+  class ClassInstance < ModuleInstance
+    def initialize(klass, superclass)
+      super(klass)
+      @superclass = superclass
+    end
+
+    def superclass
+      @superclass
+    end
+
+    # Class instance methods
+    def new
+      ObjectInstance.new(self)
+    end
+  end
+end
